web_thread_select

Trait Post

pub trait Post: Send + 'static { }

Types that can be sent to another thread. In this shim, this trait is just an alias for Send + 'static, but in web-thread some types can be sent only by performing an explicit transfer operation.